Group functions for creating exported name file together.
