+++ src/client/utils/download/download.ts
@@ -17,6 +17,13 @@
 
 import * as filesaver from "file-saver";
 import { Dataset, DatasetJSFull, TabulatorOptions } from "plywood";
+import { FixedTimeFilterClause } from "../../../common/models/filter-clause/filter-clause";
+import { Filter } from "../../../common/models/filter/filter";
+import { Split } from "../../../common/models/split/split";
+import { Splits } from "../../../common/models/splits/splits";
+import { complement } from "../../../common/utils/functional/functional";
+import { isBlank } from "../../../common/utils/general/general";
+import { STRINGS } from "../../config/constants";
 import { DataSetWithTabOptions } from "../../views/cube-view/cube-view";
 
 export type FileFormat = "csv" | "tsv" | "json";
@@ -51,9 +58,33 @@ export function datasetToFileString(dataset: Dataset, fileFormat: FileFormat, op
   }
 }
 
+function dateToFileString(date: Date): string {
+  return date.toISOString()
+    .replace("T", "_")
+    .replace("Z", "")
+    .replace(".000", "");
+}
+
+export function filter2NameComponent(filter: Filter): string {
+  const timeFilter: FixedTimeFilterClause = filter.clauses.find(clause => clause instanceof FixedTimeFilterClause) as FixedTimeFilterClause;
+  const nonTimeClauseSize = filter.clauses.filter(clause => !(clause instanceof FixedTimeFilterClause)).count();
+  const filtersPart = nonTimeClauseSize === 0 ? "" : `_filters-${nonTimeClauseSize}`;
+  if (timeFilter) {
+    const { start, end } = timeFilter.values.first();
+    return `${dateToFileString(start)}_${dateToFileString(end)}${filtersPart}`;
+  }
+  return filtersPart;
+}
+
+const split2NameComponent = (split: Split) => `${STRINGS.splitDelimiter}_${split.reference}`;
+
+export function splits2NameComponent(splits: Splits): string {
+  return splits.splits.toArray().map(split2NameComponent).join("_");
+}
+
 export function makeFileName(...nameComponents: string[]): string {
   return nameComponents
-    .filter(name => name && name.length)
+    .filter(complement(isBlank))
     .map(name => name.toLowerCase())
     .join("_")
     .substr(0, 200);
